requirements.txt
# OCR和文档处理
paddlepaddle>=2.4.0
paddleocr>=2.7.0
pdf2image>=1.16.0
Pillow>=9.0.0
opencv-python>=4.5.0
pytesseract>=0.3.10

# 机器学习和深度学习
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0
numpy>=1.21.0

# 数据处理
pandas>=1.3.0

# 其他工具
tqdm>=4.62.0
requests>=2.25.0